+++ iphone/TitaniumKit/TitaniumKit/Sources/API/ObjcProxy.m
@@ -111,7 +111,7 @@
   }
   @finally {
     pthread_rwlock_unlock(&_listenerLock);
-    [self _listenerAdded:name count:ourCallbackCount];
+    [self _listenerAdded:name count:(int)ourCallbackCount];
   }
 }
 
@@ -147,7 +147,7 @@
   @finally {
     pthread_rwlock_unlock(&_listenerLock);
     if (removed) {
-      [self _listenerRemoved:name count:ourCallbackCount];
+      [self _listenerRemoved:name count:(int)ourCallbackCount];
     }
   }
 }
